refactor: expr_builder -> operator
