+++ udon_decompiler/__init__.py
@@ -7,7 +7,6 @@ from udon_decompiler.analysis import (
     CFGBuilder,
     ControlFlowGraph,
     DataFlowAnalyzer,
-    ExpressionType,
     FunctionDataFlowAnalyzer,
     IRAssignmentStatement,
     IRBuilder,
@@ -71,7 +70,6 @@ __all__ = [
     "CFGBuilder",
     "ControlFlowGraph",
     "DataFlowAnalyzer",
-    "ExpressionType",
     "FunctionDataFlowAnalyzer",
     "StackFrame",
     "IRBuilder",
